// src/app/api-generated/model/registerReqDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterReqDto { 
    username: string;
    password: string;
    email: string;
    firstName: string;
    lastName: string;
    phone?: string;
    pfp?: string;
}